simple_one/pyproject.toml
[project]
name = "simple-one"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.26",
    "langchain-groq>=0.3.6",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.4",
    "mcp>=1.12.1",
    "python-dotenv>=1.1.1",
]
